backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Abhishek <abhishekaryz@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.1.4"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
social-auth-app-django = "^5.4.2"
python-decouple = "^3.8"
gunicorn = "^23.0.0"
whitenoise = "^6.8.2"
drf-yasg = "^1.21.8"
boto3 = "^1.35.90"
django-storages = "^1.14.4"


[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
flake8 = "^7.1.1"
pytest = "^8.3.4"
pytest-django = "^4.9.0"
pre-commit = "^4.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
